backend/requirements.txt
Flask==3.1.1
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
packaging==24.2
pinecone==7.3.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
python-dateutil==2.9.0.post0
requests==2.32.4
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
blinker==1.9.0
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
cffi==1.17.1
cryptography==45.0.6
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.22
pypdfium2==4.30.0
annotated-types==0.7.0
anyio==4.10.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9
langsmith==0.4.14
orjson==3.11.2
pydantic==2.11.7
pydantic_core==2.33.2
PyYAML==6.0.2
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.43
tenacity==9.1.2
typing-inspection==0.4.1
zstandard==0.23.0
openai==1.56.2
flask-cors==5.0.0
python-dotenv==1.0.0
pytest==8.3.3
pytest-mock==3.14.0